Split connection library into separate source files
The connection library is refactored so that the code for different
connection types live in their separate source files. In
particular, the source code for mock connections is now moved to
`test/src/net`.
