+++ src/net/conn.h
@@ -8,9 +8,8 @@ typedef enum ConnectionType
 {
 	CONNECTION_PLAIN,
 	CONNECTION_SSL,
-#ifdef DEBUG
 	CONNECTION_MOCK,
-#endif
+	_CONNECTION_MAX,
 } ConnectionType;
 
 typedef struct Connection
@@ -26,11 +25,6 @@ extern ssize_t connection_read(Connection *conn, char *buf, size_t buflen);
 extern ssize_t connection_write(Connection *conn, const char *buf, size_t writelen);
 extern void connection_close(Connection *conn);
 extern void connection_destroy(Connection *conn);
-
-#ifdef DEBUG
-/*  Special functions for a connection mocker */
-extern ssize_t connection_mock_set_recv_buf(Connection *conn, char *buf, size_t buflen);
-#endif
 
 /*  Called in init.c */
 extern void _connection_init(void);
